Add global-mode to use-package example config
At least in my setup (spacemacs 0.300 on emacs 26.1), I could not reproduce the effects of the first config example without use-package if I just picked the provided use-package alternative. Then I noticed that they are not exactly the same.

Adding `solaire-global-mode` to the example so that others do not get confused with the same mistake.