Fix previewing themes via minibuffer plugins
When you preview themes with ivy, helm, vertico, etc,
solaire-global-mode is getting toggled on the fly, which triggers
solaire-mode-fix-minibuffer, which kills certain minibuffer buffers when
a theme doesn't support solaire-mode. Doing so on Emacs 28.x seems to
cause unexplained freezes and hangs, so let's not do that while the
minibuffer is active.