fix: heuristic for detecting theme support
Before, I'd compare (face-background 'default) to (face-background
'solaire-default-face) to detect if a theme supported solaire-mode (i.e.
if they were different = yes), but this isn't reliable in daemon
sessions because there may not be an active frame available when the
user loads their theme, so no theme would pass this test.

Instead, I change it to check if the active theme has any face spec at
all for `solaire-default-face`.

Fix: #46