fix: map-y-or-n-p advice triggering too aggressively
Given the right arguments, map-y-or-n-p won't actually prompt the user.
Without a prompt, the issue this advice tries to address won't appear,
causing it to randomly flash the minibuffer's background
instead (particularly in calls to save-some-buffers, which will call
map-y-or-n-p without prompting in some scenarios).  This can be slow at
worst and distracting at best.

Fix: #50